Merge remote-tracking branch 'mainline/develop' into MAGETWO-21666-empty-array-argument

Conflicts:
	dev/tests/unit/testsuite/Magento/Framework/Data/Argument/Interpreter/ArrayTypeTest.php
	dev/tests/unit/testsuite/Magento/Framework/Data/Argument/_files/typesInvalidArray.php

+++ app/code/Magento/Reward/Model/Action/OrderExtra.php
@@ -2,8 +2,6 @@
 /**
  * {license_notice}
  *
- * @category    Magento
- * @package     Magento_Reward
  * @copyright   {copyright}
  * @license     {license_link}
  */
@@ -41,10 +39,8 @@ class OrderExtra extends \Magento\Reward\Model\Action\AbstractAction
      * @param \Magento\Reward\Helper\Data $rewardData
      * @param array $data
      */
-    public function __construct(
-        \Magento\Reward\Helper\Data $rewardData,
-        array $data = array()
-    ) {
+    public function __construct(\Magento\Reward\Helper\Data $rewardData, array $data = array())
+    {
         $this->_rewardData = $rewardData;
         parent::__construct($data);
     }
@@ -64,15 +60,13 @@ class OrderExtra extends \Magento\Reward\Model\Action\AbstractAction
     /**
      * Setter for $_entity and add some extra data to history
      *
-     * @param \Magento\Object $entity
+     * @param \Magento\Framework\Object $entity
      * @return $this
      */
     public function setEntity($entity)
     {
         parent::setEntity($entity);
-        $this->getHistory()->addAdditionalData(array(
-            'increment_id' => $this->getEntity()->getIncrementId()
-        ));
+        $this->getHistory()->addAdditionalData(array('increment_id' => $this->getEntity()->getIncrementId()));
         return $this;
     }
 
@@ -104,16 +98,16 @@ class OrderExtra extends \Magento\Reward\Model\Action\AbstractAction
             // known issue: no support for multishipping quote
             $address = $quote->getIsVirtual() ? $quote->getBillingAddress() : $quote->getShippingAddress();
             // use only money customer spend - shipping & tax
-            $monetaryAmount = $quote->getBaseGrandTotal()
-                - $address->getBaseShippingAmount()
-                - $address->getBaseTaxAmount();
+            $monetaryAmount = $quote->getBaseGrandTotal() -
+                $address->getBaseShippingAmount() -
+                $address->getBaseTaxAmount();
             $monetaryAmount = $monetaryAmount < 0 ? 0 : $monetaryAmount;
         } else {
-            $monetaryAmount = $this->getEntity()->getBaseTotalPaid()
-                - $this->getEntity()->getBaseShippingAmount()
-                - $this->getEntity()->getBaseTaxAmount();
+            $monetaryAmount = $this->getEntity()->getBaseTotalPaid() -
+                $this->getEntity()->getBaseShippingAmount() -
+                $this->getEntity()->getBaseTaxAmount();
         }
-        $pointsDelta = $this->getReward()->getRateToPoints()->calculateToPoints((float)$monetaryAmount);
+        $pointsDelta = $this->getReward()->getRateToPoints()->calculateToPoints((double)$monetaryAmount);
         return $pointsDelta;
     }
 
@@ -125,7 +119,6 @@ class OrderExtra extends \Magento\Reward\Model\Action\AbstractAction
      */
     public function canAddRewardPoints()
     {
-        return parent::canAddRewardPoints()
-            && $this->_rewardData->isOrderAllowed($this->getReward()->getWebsiteId());
+        return parent::canAddRewardPoints() && $this->_rewardData->isOrderAllowed($this->getReward()->getWebsiteId());
     }
 }
