+++ app/code/Magento/Sales/Model/Order/Invoice/Total/Shipping.php
@@ -2,8 +2,6 @@
 /**
  * {license_notice}
  *
- * @category    Magento
- * @package     Magento_Sales
  * @copyright   {copyright}
  * @license     {license_link}
  */
@@ -12,8 +10,6 @@ namespace Magento\Sales\Model\Order\Invoice\Total;
 /**
  * Order invoice shipping total calculation model
  *
- * @category   Magento
- * @package    Magento_Sales
  * @author      Magento Core Team <core@magentocommerce.com>
  */
 class Shipping extends AbstractTotal
@@ -26,16 +22,16 @@ class Shipping extends AbstractTotal
     {
         $invoice->setShippingAmount(0);
         $invoice->setBaseShippingAmount(0);
-        $orderShippingAmount        = $invoice->getOrder()->getShippingAmount();
-        $baseOrderShippingAmount    = $invoice->getOrder()->getBaseShippingAmount();
-        $shippingInclTax            = $invoice->getOrder()->getShippingInclTax();
-        $baseShippingInclTax        = $invoice->getOrder()->getBaseShippingInclTax();
+        $orderShippingAmount = $invoice->getOrder()->getShippingAmount();
+        $baseOrderShippingAmount = $invoice->getOrder()->getBaseShippingAmount();
+        $shippingInclTax = $invoice->getOrder()->getShippingInclTax();
+        $baseShippingInclTax = $invoice->getOrder()->getBaseShippingInclTax();
         if ($orderShippingAmount) {
             /**
              * Check shipping amount in previous invoices
              */
             foreach ($invoice->getOrder()->getInvoiceCollection() as $previousInvoice) {
-                if ((float)$previousInvoice->getShippingAmount() && !$previousInvoice->isCanceled()) {
+                if ((double)$previousInvoice->getShippingAmount() && !$previousInvoice->isCanceled()) {
                     return $this;
                 }
             }
@@ -44,8 +40,8 @@ class Shipping extends AbstractTotal
             $invoice->setShippingInclTax($shippingInclTax);
             $invoice->setBaseShippingInclTax($baseShippingInclTax);
 
-            $invoice->setGrandTotal($invoice->getGrandTotal()+$orderShippingAmount);
-            $invoice->setBaseGrandTotal($invoice->getBaseGrandTotal()+$baseOrderShippingAmount);
+            $invoice->setGrandTotal($invoice->getGrandTotal() + $orderShippingAmount);
+            $invoice->setBaseGrandTotal($invoice->getBaseGrandTotal() + $baseOrderShippingAmount);
         }
         return $this;
     }
